refactor: convert `convert` into class
vyper/types/convert.py is moved to vyper/functions/convert.py
because the logic in this module is only used within the builtin
`convert` method.
